+++ pt/pt_config.php
@@ -4,7 +4,7 @@
  */
 
 return [
-    'version' => '4.0.4',
+    'version' => '4.1.0',
     'update_date' => '2026-07-15',
     'enabled' => true,